Add error handling
add modification feature
add error handling

+++ src/interfaces/MainInterface.java
@@ -12,6 +12,8 @@ import javax.swing.ListSelectionModel;
 
 import java.awt.Color;
 import javax.swing.JLabel;
+import javax.swing.JOptionPane;
+
 import java.awt.Font;
 import javax.swing.SwingConstants;
 import javax.swing.table.DefaultTableModel;
@@ -308,12 +310,22 @@ public class MainInterface {
 		mainPanel.add(reservations, "reserv");
 		reservations.setLayout(null);
 		
+		JLabel reserv_warning_lbl = new JLabel("");
+		reserv_warning_lbl.setForeground(Color.RED);
+		reserv_warning_lbl.setBounds(517, 57, 185, 88);
+		reservations.add(reserv_warning_lbl);
 		
 		JScrollPane reserv_scroll = new JScrollPane();
 		reserv_scroll.setBounds(23, 57, 484, 462);
 		reservations.add(reserv_scroll);
 		
-		reserv_table = new JTable();
+		reserv_table = new JTable() {
+			private static final long serialVersionUID = 1L;
+			public boolean isCellEditable(int row, int column){  
+		          return false;  
+		    };
+		};
+		
 		reserv_table.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
 		reserv_scroll.setViewportView(reserv_table);
 		ReservationController.fetchAll(reserv_table, filtre.Tous);
@@ -328,6 +340,9 @@ public class MainInterface {
 		reserv_actualiser_btn.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				ReservationController.fetchAll(reserv_table, (filtre)reserv_filtre.getSelectedItem());
+				
+				//Reset warning label on succesful reload
+				reserv_warning_lbl.setText("");
 			}
 		});
 		reserv_actualiser_btn.setBounds(517, 463, 185, 56);
@@ -340,8 +355,11 @@ public class MainInterface {
 		JButton newReserv_btn = new JButton("Nouveau reservation");
 		newReserv_btn.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
+				//Open reservation creation window
 				CreerReservation newReserv = new CreerReservation(reserv_table);
-				ReservationController.fetchAll(reserv_table, (filtre)reserv_filtre.getSelectedItem());
+				
+				//Reset warning label on succesful operation
+				reserv_warning_lbl.setText("");
 			}
 		});
 		newReserv_btn.setBounds(517, 155, 185, 56);
@@ -350,16 +368,55 @@ public class MainInterface {
 		JButton delReserv_btn = new JButton("Supprimer reservation");
 		delReserv_btn.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
+				
 				int index = reserv_table.getSelectedRow();
-				String codeReserv = (String) reserv_table.getValueAt(index, 0);
-				ReservationController.deleteReservation(codeReserv);
-				ReservationController.fetchAll(reserv_table, (filtre)reserv_filtre.getSelectedItem());
+				if(index < 0) {
+					// if user hasnt selected any row :
+					reserv_warning_lbl.setText("<html>Veuillez Selectionner une reservation à supprimer.</html>");
+					// ^ html tag is for automatic text wrapping
+					return;
+				} //else
+				
+				
+				// Verification prompt, ask user if he really wants to delete element
+				int result = JOptionPane.showConfirmDialog(reservations, "Etes vous sûr?","Verification", JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE);
+				
+				//if user wants to delete element
+				if(result == JOptionPane.YES_OPTION) {
+					String codeReserv = (String) reserv_table.getValueAt(index, 0);
+					ReservationController.deleteReservation(codeReserv);
+					ReservationController.fetchAll(reserv_table, (filtre)reserv_filtre.getSelectedItem());	
+				}
+				//Reset warning label on succesful operation
+				reserv_warning_lbl.setText("");
 			}
 		});
 		delReserv_btn.setBounds(517, 225, 185, 56);
 		reservations.add(delReserv_btn);
 		
 		JButton modReserv_btn = new JButton("Modifier reservation");
+		modReserv_btn.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent e) {
+				int index = reserv_table.getSelectedRow();
+				if(index < 0) {
+					// if user hasnt selected any row :
+					reserv_warning_lbl.setText("<html>Veuillez Selectionner une reservation à modifier.</html>");
+					// ^ html tag is for automatic text wrapping
+					return;
+				}
+				String codeReserv = reserv_table.getValueAt(index, 0).toString();
+				String dateDep = reserv_table.getValueAt(index, 3).toString();
+				String dateRet = reserv_table.getValueAt(index, 4).toString();
+				boolean isValid = (boolean)reserv_table.getValueAt(index, 5);
+				boolean isCanceled = (boolean)reserv_table.getValueAt(index, 6);
+				
+				//Open reservation modification window
+				ModifierReservation newReserv = new ModifierReservation(reserv_table, codeReserv, dateDep, dateRet, isValid, isCanceled);
+				
+				//Reset warning label on succesful operation
+				reserv_warning_lbl.setText("");
+			}
+		});
 		modReserv_btn.setBounds(517, 291, 185, 56);
 		reservations.add(modReserv_btn);
 		
@@ -373,10 +430,17 @@ public class MainInterface {
 		searchReserv_btn.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				String codeReserv = reserv_field.getText();
+				if(codeReserv.isEmpty()) {
+					reserv_warning_lbl.setText("<html>Veuillez Saisir un code à rechercher.</html>");
+					return;
+				}
+				//search for reservation
 				ReservationController.findReservation(codeReserv, reserv_table);
+				
+				//Reset warning label on succesful operation
+				reserv_warning_lbl.setText("");
 			}
 		});
 		reservations.add(searchReserv_btn);
-		
 	}
 }
